zlogin primarily for osx specifics

+++ .zlogin
@@ -1,11 +1,14 @@
-unalias ps
-alias brew_update='brew update && brew upgrade --all && brew cleanup && brew services cleanup && brew doctor && brew linkapps && brew prune'
-alias password='ruby ~/tools/password.rb -an'
-alias r='bin/rails'
-alias rspec='rspec --color spec'
+if [[ $_system_name == "OSX" ]];
+then
+    unalias ps
+    alias brew_update='brew update && brew upgrade --all && brew cleanup && brew services cleanup && brew doctor && brew linkapps && brew prune'
+    alias password='ruby ~/tools/password.rb -an'
+    alias r='bin/rails'
+    alias rspec='rspec --color spec'
 
-# recursive, preserve times, verbose, update (only if change time greater)
-alias sync_music='rsync --iconv=utf8 -rtvu --delete "/Users/kirk/Music/iTunes/iTunes Media/Music/" /Volumes/xtc/Media'
+    # recursive, preserve times, verbose, update (only if change time greater)
+    alias sync_music='rsync --iconv=utf8 -rtvu --delete "/Users/kirk/Music/iTunes/iTunes Media/Music/" /Volumes/xtc/Media'
+fi
 
 # less colors
 export CLICOLOR_FORCE=1
